+++ package.js
@@ -15,6 +15,6 @@ Package.on_use(function(api) {
     api.export(['AutoForm'], ['client', 'server']);
   }
 
-  api.add_files(['autoform-common.js'], ['client', 'server']);
+  api.add_files(['autoform-server.js'], ['server']);
   api.add_files(['autoform.html', 'autoform-client.js'], ['client']);
 });