+++ CHANGELOG.md
@@ -6,6 +6,42 @@ forms with automatic insert and update events, and automatic reactive validation
 
 ## Change Log
 
+### 0.4.4
+
+* Fix issues, introduced by 0.4.3, where select, checkbox, and radio form values
+were not always correct.
+* Throw errors in afFieldMessage and afFieldIsInvalid helpers when the field
+name used doesn't exist in the schema. (Other helpers already did this.)
+* Add two new "type" options for quickForms: "readonly" and "disabled". See
+README.
+* When displaying a quickForm, use the allowedValues from the schema as the
+options.
+
+### 0.4.3
+
+Update to use MongoObject to create flatDoc. Fixes an issue caused by recent
+SimpleSchema API changes.
+
+### 0.4.2
+
+* New API. The old API is deprecated but continues to work for now with warning
+messages logged. See the README and warning messages. The main difference is
+that the various hooks/callbacks are centralized in a `hooks()` method, but this
+is available only on `AutoForm` instances, so you need to wrap your `Collection2`
+instance in an `AutoForm` on the client and then attach hooks to the autoform,
+passing the autoform as the `schema` attribute for your `autoForm` helper. Also,
+the server methods have all been removed, making this a purely client package.
+This means you should now define a `SimpleSchema` in common code and use that
+to create an `AutoForm` instance on the client only. This allows you to still
+validate against the same `SimpleSchema` on both the client and the server. See
+examples in the README.
+* You can now set `options` to the string "allowed" for `afFieldInput` or
+`afQuickField`. This causes the schema's `allowedValues` to be used as the
+values (and labels) for the select element. If you want to capitalize the first
+letter of the labels, set `capitalize="true"`. (Thanks @gdhuse!)
+* SimpleSchema error messages with HTML will now display correctly when you use
+an `afQuickField`. (Thanks @gdhuse!)
+
 ### 0.4.1
 
 Improve handling of key names that use dot notation to indicate a property of